// src/app/services/auth/models/user-dto.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

export interface UserDto {
  avatarUrl?: string;
  createdAt?: string;
  email?: string;
  emailVerified?: boolean;
  firstName?: string;
  fullName?: string;
  id?: string;
  isActive?: boolean;
  lastLoginAt?: string;
  lastName?: string;
  roles?: Array<string>;
}